requirements.txt
django==4.1.2
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0
